refactor: merge internal/shared/version into internal/project
